comment-crud/src/main/java/com/cognizant/commentcrud/controller/CommentController.java
package com.cognizant.commentcrud.controller;

import com.cognizant.commentcrud.model.Comment;
import com.cognizant.commentcrud.service.CommentService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/comments")
public class CommentController {

    private CommentService service;

    @Autowired
    public CommentController(CommentService service) {
        this.service = service;
    }

    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    public Page<Comment> getAllCommentsByPostId(
            @RequestParam(defaultValue = "1") Integer postId,
            @RequestParam(defaultValue = "0") Integer pageNumber,
            @RequestParam(defaultValue = "5") Integer pageSize
    ) {
        return service.getAllCommentsByPostId(postId, pageNumber, pageSize);
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public Comment createComment(@RequestBody Comment comment){
        return service.createComment(comment);
    }

}
